@@ -1,20 +1,20 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-    <PropertyGroup>
-      <TargetFramework>net6.0</TargetFramework>
-      <OutputType>Exe</OutputType>
-      <IsPackable>false</IsPackable>
-    </PropertyGroup>
+  <PropertyGroup>
+    <TargetFramework>net7.0</TargetFramework>
+    <OutputType>Exe</OutputType>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <LangVersion>latest</LangVersion>
+    <IsPackable>false</IsPackable>
+  </PropertyGroup>
 
-    <ItemGroup>
-      <ProjectReference Include="..\AspNetCoreExample.Api\AspNetCoreExample.Api.csproj" />
-      <ProjectReference Include="..\TypeScript.ContractGenerator\TypeScript.ContractGenerator.csproj" />
-    </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\AspNetCoreExample.Api\AspNetCoreExample.Api.csproj" />
+  </ItemGroup>
 
-    <ItemGroup>
-      <TypeScriptCompile Include="output\apiBase\ApiBase.ts" />
-      <TypeScriptCompile Include="output\apiBase\UserApiBase.ts" />
-      <TypeScriptCompile Include="output\dataTypes\Guid.ts" />
-    </ItemGroup>
+  <ItemGroup>
+    <TypeScriptCompile Include="output\**\*.ts" />
+  </ItemGroup>
 
 </Project>
