@@ -1,10 +1,7 @@
-using System;
+﻿namespace AspNetCoreExample.Api.Models;
 
-namespace AspNetCoreExample.Api.Models
+public class User
 {
-    public class User
-    {
-        public Guid Id { get; set; }
-        public string Name { get; set; }
-    }
+    public Guid Id { get; set; }
+    public string Name { get; set; }
 }
