@@ -1,12 +1,12 @@
 // tslint:disable
 // TypeScriptContractGenerator's generated content
-import { User } from './../dto/User';
-import { Guid } from './../dataTypes/Guid';
-import { ApiBase } from './../apiBase/ApiBase';
+import { User } from './../DataTypes/User';
+import { Guid } from './../DataTypes/Guid';
+import { ApiBase } from './../ApiBase/ApiBase';
 
 export class UserApi extends ApiBase implements IUserApi {
     async createUser(user: User): Promise<void> {
-        return this.makePostRequest(`user/`, {
+        return this.makePostRequest(`/v1/users`, {
             
         }, {
             ...user,
@@ -14,7 +14,7 @@ export class UserApi extends ApiBase implements IUserApi {
     }
 
     async deleteUser(userId: Guid): Promise<void> {
-        return this.makeDeleteRequest(`user/${userId}`, {
+        return this.makeDeleteRequest(`/v1/users/${userId}`, {
             
         }, {
             
@@ -22,8 +22,16 @@ export class UserApi extends ApiBase implements IUserApi {
     }
 
     async getUser(userId: Guid): Promise<User> {
-        return this.makeGetRequest(`user/${userId}`, {
+        return this.makeGetRequest(`/v1/users/${userId}`, {
             
+        }, {
+            
+        });
+    }
+
+    async searchUsers(name: string): Promise<User[]> {
+        return this.makeGetRequest(`/v1/users`, {
+            ['name']: name,
         }, {
             
         });
@@ -34,4 +42,5 @@ export interface IUserApi {
     createUser(user: User): Promise<void>;
     deleteUser(userId: Guid): Promise<void>;
     getUser(userId: Guid): Promise<User>;
+    searchUsers(name: string): Promise<User[]>;
 }
