@@ -1,6 +1,6 @@
 // tslint:disable
 // TypeScriptContractGenerator's generated content
-import { Guid } from './../dataTypes/Guid';
+import { Guid } from './Guid';
 
 export type User = {
     id: Guid;
